Moved 'MIPSInstruction' to the header
